requirements.txt
pandas
polars
requests
pypdf
streamlit
langchain
pygwalker
cryptography
ollama
llama-index-llms-ollama